# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobAccepted": {
        "properties": {
          "cost_credits": {
            "title": "Cost Credits",
            "type": "integer"
          },
          "estimated_wait_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Seconds"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "cost_credits"
        ],
        "title": "JobAccepted",
        "type": "object"
      },
      "JobOut": {
        "properties": {
          "cost_credits": {
            "title": "Cost Credits",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "error_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Stage"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "negative_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negative Prompt"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "result_image_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Result Image Ids",
            "type": "array"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "stages": {
            "items": {},
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "progress",
          "cost_credits",
          "prompt",
          "negative_prompt",
          "seed",
          "stages",
          "params",
          "error_stage",
          "error_code",
          "error_message",
          "priority",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "JobOut",
        "type": "object"
      },
      "JobPage": {
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "JobPage",
        "type": "object"
      },
      "JobParams": {
        "properties": {
          "guidance": {
            "default": 3.5,
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Guidance",
            "type": "number"
          },
          "height": {
            "default": 1024,
            "maximum": 1536.0,
            "minimum": 256.0,
            "title": "Height",
            "type": "integer"
          },
          "num_outputs": {
            "default": 1,
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Num Outputs",
            "type": "integer"
          },
          "steps": {
            "default": 28,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Steps",
            "type": "integer"
          },
          "width": {
            "default": 1024,
            "maximum": 1536.0,
            "minimum": 256.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "title": "JobParams",
        "type": "object"
      },
      "JobSubmitRequest": {
        "properties": {
          "negative_prompt": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negative Prompt"
          },
          "params": {
            "$ref": "#/components/schemas/JobParams"
          },
          "prompt": {
            "maxLength": 4000,
            "title": "Prompt",
            "type": "string"
          },
          "reference_image_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Reference Image Ids",
            "type": "array"
          },
          "seed": {
            "anyOf": [
              {
                "maximum": 2147483647.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "stages": {
            "items": {
              "type": "string"
            },
            "title": "Stages",
            "type": "array"
          },
          "style_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Slug"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "JobSubmitRequest",
        "type": "object"
      },
...
      "PackItem": {
        "properties": {
          "cost_credits": {
            "title": "Cost Credits",
            "type": "integer"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "style_slug": {
            "title": "Style Slug",
            "type": "string"
          },
          "variant": {
            "title": "Variant",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "style_slug",
          "variant",
          "cost_credits",
          "status"
        ],
        "title": "PackItem",
        "type": "object"
      },
      "PackRequest": {
        "description": "Avatar pack: generate across multiple styles (and variants) in one request.",
        "properties": {
          "negative_prompt": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negative Prompt"
          },
          "params": {
            "$ref": "#/components/schemas/JobParams"
          },
          "prompt": {
            "maxLength": 4000,
            "title": "Prompt",
            "type": "string"
          },
          "reference_image_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Reference Image Ids",
            "type": "array"
          },
          "stages": {
            "items": {
              "type": "string"
            },
            "title": "Stages",
            "type": "array"
          },
          "style_slugs": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Style Slugs",
            "type": "array"
          },
          "variants_per_style": {
            "default": 1,
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Variants Per Style",
            "type": "integer"
          }
        },
        "required": [
          "prompt",
          "style_slugs"
        ],
        "title": "PackRequest",
        "type": "object"
      },
      "PackResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/PackItem"
            },
            "title": "Jobs",
            "type": "array"
          },
          "pack_id": {
            "title": "Pack Id",
            "type": "string"
          },
          "total_cost_credits": {
            "title": "Total Cost Credits",
            "type": "integer"
          }
        },
        "required": [
          "pack_id",
          "total_cost_credits",
          "jobs"
        ],
        "title": "PackResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API-first, self-hosted AI image-generation platform. All product capabilities are versioned HTTP endpoints under /api/v1.",
    "title": "AI Mirror Platform",
    "version": "1.0.0",
    "x-error-envelope": {
      "error": {
        "code": "string",
        "details": {},
        "message": "string"
      }
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List your jobs",
        "tags": [
          "generation"
        ]
      },
      "post": {
        "operationId": "submit_job_api_v1_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit a generation job",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/jobs/pack": {
      "post": {
        "operationId": "submit_pack_api_v1_jobs_pack_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit an avatar pack (multiple styles/variants)",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job status / progress",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel a job",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/events": {
      "get": {
        "operationId": "job_events_api_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "SSE progress stream",
        "tags": [
          "generation"
        ]
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}